Mark Snap as grade: stable
Snap packages marked as `grade: devel` are not permitted to be released
to the `candidate` or `stable` channels in the Snap Store. This commit
changes the grade to `stable` now that everything seems to be working
well. Snaps build from this commit forward will be releasable to the
`candidate` and/or `stable` channels in the Snap Store :-)

Signed-off-by: Daniel Llewellyn <daniel@snapcraft.ninja>